+++ compiler/vmgen.nim
@@ -1366,6 +1366,8 @@ proc genMagic(c: PCtx; n: PNode; dest: var TDest; m: TMagic) =
     c.freeTemp(a)
   of mNodeId:
     c.genUnaryABC(n, dest, opcNodeId)
+  of mNIsUnion:
+    c.genUnaryABC(n, dest, opcNIsUnion)
   else:
     # mGCref, mGCunref,
     globalError(c.config, n.info, "cannot generate code for: " & $m)